clients/cli/pyproject.toml
[project]
name = "llmchat-cli-client"
version = "1.0.0"
description = "LLMChatAssistant CLI客户端"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["cli"]

[tool.setuptools.package-dir]
cli = ""
